lsp: Move `LangugeFeature` to `esbonio.lsp.rst`
This resolves the "more than one target found for cross-reference"
error when building the docs.
