--- docs/contributing/lsp/feature.rst
@@ -1,4 +1,0 @@
-Feature
-=======
-
-.. automodule:: esbonio.lsp.feature
